Move src/ApolloClient.ts into src/core/, and update tests.
This allows src/core/index.ts to avoid reaching outside the
@apollo/client/core entry point to import the all-important ApolloClient
module/class.
