+++ src/__tests__/ApolloClient.ts
@@ -1,14 +1,18 @@
 import gql from 'graphql-tag';
 
-import { Observable } from '../utilities/observables/Observable';
-import { makeReference } from '../core';
-import { ApolloLink } from '../link/core/ApolloLink';
-import { HttpLink } from '../link/http/HttpLink';
-import { InMemoryCache } from '../cache/inmemory/inMemoryCache';
-import { stripSymbols } from '../utilities/testing/stripSymbols';
-import { ApolloClient } from '../';
-import { DefaultOptions } from '../ApolloClient';
-import { FetchPolicy, QueryOptions } from '../core/watchQueryOptions';
+import {
+  ApolloClient,
+  DefaultOptions,
+  FetchPolicy,
+  QueryOptions,
+  makeReference,
+} from '../core';
+
+import { Observable } from '../utilities';
+import { ApolloLink } from '../link/core';
+import { HttpLink } from '../link/http';
+import { InMemoryCache } from '../cache';
+import { stripSymbols } from '../testing';
 
 describe('ApolloClient', () => {
   describe('constructor', () => {
