+++ src/react/components/__tests__/ssr/server.test.tsx
@@ -10,12 +10,12 @@ import {
 } from 'graphql';
 import gql from 'graphql-tag';
 
-import { ApolloClient } from '../../../../ApolloClient';
-import { InMemoryCache as Cache } from '../../../../cache/inmemory/inMemoryCache';
-import { ApolloProvider } from '../../../context/ApolloProvider';
+import { ApolloClient } from '../../../../core';
+import { InMemoryCache as Cache } from '../../../../cache';
+import { ApolloProvider } from '../../../context';
 import { ApolloLink } from '../../../../link/core';
-import { Observable } from '../../../../utilities/observables/Observable';
-import { renderToStringWithData } from '../../../ssr/renderToStringWithData';
+import { Observable } from '../../../../utilities';
+import { renderToStringWithData } from '../../../ssr';
 import { Query } from '../../Query';
 
 const planetMap = new Map([['Planet:1', { id: 'Planet:1', name: 'Tatooine' }]]);
