+++ src/react/hooks/__tests__/useApolloClient.test.tsx
@@ -2,12 +2,11 @@ import React from 'react';
 import { render, cleanup } from '@testing-library/react';
 import { InvariantError } from 'ts-invariant';
 
+import { ApolloClient } from '../../../core';
 import { ApolloLink } from '../../../link/core';
-import { ApolloProvider } from '../../context/ApolloProvider';
-import { ApolloClient } from '../../../ApolloClient';
-import { InMemoryCache } from '../../../cache/inmemory/inMemoryCache';
+import { ApolloProvider, resetApolloContext } from '../../context';
+import { InMemoryCache } from '../../../cache';
 import { useApolloClient } from '../useApolloClient';
-import { resetApolloContext } from '../../context/ApolloContext';
 
 describe('useApolloClient Hook', () => {
   afterEach(() => {
